requirements.txt
click==7.1.2
dominate==2.5.1
Flask==1.1.2
Flask-Bootstrap4==4.0.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
SQLAlchemy==1.3.18
visitor==0.1.3
Werkzeug==1.0.1
